Add Tax Invoice Operations

+++ src/SageOneApi/Models/AdditionalItemPrice.cs
@@ -3,8 +3,8 @@ namespace SageOneApi.Models
 	public class AdditionalItemPrice : BaseModel
 	{
 		public int ItemId { get; set; }
-		public float PriceInclusive { get; set; }
-		public float PriceExclusive { get; set; }
+		public decimal PriceInclusive { get; set; }
+		public decimal PriceExclusive { get; set; }
 		public int AdditionalPriceListId { get; set; }
 	}
 }
