+++ src/SageOneApi/Models/TaxType.cs
@@ -3,7 +3,7 @@
 	public class TaxType : BaseModel
 	{
 		public string Name { get; set; }
-		public float Percentage { get; set; }
+		public decimal Percentage { get; set; }
 		public bool IsDefault { get; set; }
 		public bool HasActivity { get; set; }
 		public bool IsManualTax { get; set; }
